refactor: MainPage 훅 디렉토리 수정 #276
