+++ frontend/src/pages/MainPage/hooks/useCreateRoom.ts
@@ -2,11 +2,14 @@ import { useNavigate } from 'react-router-dom';
 
 import { ROUTES } from '@/constants/routes';
 
-export const useCreateRoom = () => {
+const useCreateRoom = () => {
   const navigate = useNavigate();
+
   const handleRoomCreate = () => {
     navigate(ROUTES.nickname);
   };
 
   return { handleRoomCreate };
 };
+
+export default useCreateRoom;
